requirements.txt
djangorestframework
djangorestframework-simplejwt 
drf-yasg
django-filter
setuptools